requirements.txt
azure-functions>=1.2.0
#-e git+https://github.com/Azure/azure-functions-durable-python.git@f1460c3534d11952a5b44584a76f93a084c7e78a#egg=azure_functions_durable